+++ src/main.cpp
@@ -65,11 +65,11 @@ mainClass(const char* jar)
    public:
     MyAllocator(System* s): s(s) { }
 
-    virtual void* tryAllocate(unsigned size) {
+    virtual void* tryAllocate(size_t size) {
       return s->tryAllocate(size);
     }
 
-    virtual void* allocate(unsigned size) {
+    virtual void* allocate(size_t size) {
       void* p = tryAllocate(size);
       if (p == 0) {
         abort(s);
@@ -77,7 +77,7 @@ mainClass(const char* jar)
       return p;
     }
 
-    virtual void free(const void* p, unsigned) {
+    virtual void free(const void* p, size_t) {
       s->free(p);
     }
